Fix unfinished tutorial hiding
